Add Ui class. Work on Storage Class
