+++ src/main/java/Scanner.java
@@ -1,0 +1,4 @@
+
+public class Scanner {
+
+}
